+++ huey/storage.py
@@ -15,6 +15,11 @@ import threading
 import time
 
 try:
+    import cysqlite
+except ImportError:
+    cysqlite = None
+
+try:
     from redis import ConnectionPool
     from redis import Redis
     from redis.exceptions import ConnectionError
@@ -833,7 +838,7 @@ class SqliteStorage(BaseSqlStorage):
            table_counter]
 
     def __init__(self, name='huey', filename='huey.db', cache_mb=8,
-                 fsync=False, journal_mode='wal', timeout=5, strict_fifo=False,
+                 fsync=None, journal_mode='wal', timeout=5, strict_fifo=False,
                  create_tables=True, **kwargs):
         self.filename = filename
         self._cache_mb = cache_mb
@@ -855,7 +860,7 @@ class SqliteStorage(BaseSqlStorage):
                 'primary key autoincrement')
             self.ddl = ddl
 
-        self.to_blob = lambda b: sqlite3.Binary(b)
+        self.to_blob = memoryview
 
         super(SqliteStorage, self).__init__(name, create_tables=create_tables)
 
@@ -1022,6 +1027,46 @@ class SqliteStorage(BaseSqlStorage):
 
     def flush_counters(self):
         self.sql('delete from counter where queue=?', (self.name,), True)
+
+
+class CySqliteStorage(SqliteStorage):
+    def __init__(self, name='huey', filename='huey.db', pragmas=None,
+                 strict_fifo=False, create_tables=True, **kwargs):
+        if cysqlite is None:
+            raise ConfigurationError('"cysqlite" not found. Run "pip install '
+                                     'cysqlite" to install.')
+        # Normalize hard-coded params to generic pragmas.
+        pragmas = pragmas or {}
+        pragmas.setdefault('journal_mode', 'wal')
+        if 'journal_mode' in kwargs:
+            pragmas['journal_mode'] = kwargs.pop('journal_mode') or 'wal'
+        if 'cache_mb' in kwargs:
+            pragmas['cache_size'] = kwargs.pop('cache_mb') * -1000
+        if 'fsync' in kwargs:
+            pragmas['synchronous'] = 2 if kwargs.pop('synchronous') else 0
+
+        super(CySqliteStorage, self).__init__(
+            name,
+            filename,
+            strict_fifo=strict_fifo,
+            create_tables=create_tables,
+            pragmas=pragmas,
+            **kwargs)
+
+    def _create_connection(self):
+        return cysqlite.connect(self.filename, timeout=self._timeout,
+                                **self._conn_kwargs)
+
+    def put_if_empty(self, key, value):
+        try:
+            with self.db(commit=True) as curs:
+                curs.execute('insert or abort into kv '
+                             '(queue, key, value) values (?, ?, ?)',
+                             (self.name, key, self.to_blob(value)))
+        except cysqlite.IntegrityError:
+            return False
+        else:
+            return True
 
 
 class PostgresStorage(BaseSqlStorage):